fix merge issues
